Use a dummy container as base networking stack
There was a chicken and egg problem. The services need to be available
before the primary container, but need the network of the primary
container to become available. So this creates a dummy 'sleep forever'
container which is the base networking stack, piggy backs the services
onto this container's network, then runs the commands in primary
container.